~2% reduced flash usage

+++ dep/SSD1351-Driver-Library/src/fonts.h
@@ -25,7 +25,7 @@ typedef struct {
 	uint8_t width;    // Font width in pixels
 	uint8_t height;   // Font height in pixels
     uint8_t first;
-	uint16_t *data; // Pointer to data font data array
+	uint8_t *data; // Pointer to data font data array
 } font_t;
 
 #endif // SSD1351_FONTS_H
